+++ NEWS.md
@@ -1,3 +1,5 @@
+# testdat (development version)
+
 # testdat 0.3.0
 
 ## Breaking changes